calculateEffStat splitted
