+++ interface/src/App/PvP/components/EditPokemon.jsx
@@ -11,7 +11,7 @@ import MagicBox from "./MagicBox/MagicBox"
 import Button from "App/Components/Button/Button";
 
 import { MovePoolBuilder } from "js/movePoolBuilder";
-import { calculateEffStat } from "js/indexFunctions";
+import { calculateEffStat } from "js/calculateEffStat";
 import { calculateMaximizedStats } from "js/Maximizer/Maximizer";
 import { processInitialStats } from "js/checks/processInitialStats";
 import { checkLvl } from "js/checks/checkLvl";
