+++ interface/src/App/PvP/components/MatrixPanel.jsx
@@ -28,7 +28,8 @@ import MatrixPokemonList from "./MatrixPokemonList/MatrixPokemonList";
 import SaveMenu from "./SaveMenu/SaveMenu";
 
 import { MovePoolBuilder } from "js/movePoolBuilder";
-import { calculateEffStat, pokemon, } from "js/indexFunctions.js";
+import { pokemon, } from "js/indexFunctions.js";
+import { calculateEffStat } from "js/calculateEffStat";
 import { calculateMaximizedStats } from "js/Maximizer/Maximizer";
 import { processInitialStats } from "js/checks/processInitialStats";
 import { checkLvl } from "js/checks/checkLvl";
